Move Ptr to graphmap module
